centralize DB initialization logic into db package
