--- common/dbconfig.go
@@ -1,107 +1,0 @@
-package common
-
-import (
-	"errors"
-	"fmt"
-	"log"
-
-	"gorm.io/driver/mysql"
-	"gorm.io/driver/postgres"
-	"gorm.io/driver/sqlite"
-	"gorm.io/gorm"
-	"gorm.io/gorm/logger"
-)
-
-// DBConfig holds configuration information to connect to a database.
-// Parameters for the config.
-//   - dbname - The name of the database to connect to
-//   - user - The user to sign in as
-//   - password - The user's password
-//   - host - The host to connect to. Values that start with / are for unix domain sockets.
-//     (default is localhost)
-//   - port - The port to bind to. (default is 5432)
-//   - sslmode - Whether or not to use SSL (default is require, this is not the default for libpq)
-//     Valid SSL modes:
-//   - disable - No SSL
-//   - require - Always SSL (skip verification)
-//   - verify-full - Always SSL (require verification)
-type DBConfig struct {
-	DbType   string `yaml:"db_type" validate:"required"`
-	URL      string `yaml:"url" validate:"required"`
-	Username string `yaml:"username" validate:"required"`
-	Password string `yaml:"password" validate:"required"`
-	DbName   string `yaml:"db_name" validate:"required"`
-	Sslmode  string `yaml:"ssl_mode" validate:"required"`
-	Port     int64  `yaml:"port" validate:"required"` // Is int64 to match the type that rds.Endpoint.Port is in the AWS RDS SDK.
-}
-
-const defaultMaxDbConnections = 10
-
-// DBInit is a generic helper function that will try to connect to a database with the config in the input.
-// Supported DB types:
-// * postgres
-// * mysql
-// * sqlite3
-func DBInit(dbConfig *DBConfig) (*gorm.DB, error) {
-	var DB *gorm.DB
-	var err error
-
-	var maxDbConnections int
-	maxDbConnections = defaultMaxDbConnections
-
-	switch dbConfig.DbType {
-	case "postgres":
-		conn := "dbname=%s user=%s password=%s host=%s sslmode=%s port=%d"
-		conn = fmt.Sprintf(conn,
-			dbConfig.DbName,
-			dbConfig.Username,
-			dbConfig.Password,
-			dbConfig.URL,
-			dbConfig.Sslmode,
-			dbConfig.Port)
-		DB, err = gorm.Open(postgres.Open(conn), &gorm.Config{})
-	case "mysql":
-		conn := "%s:%s@%s(%s:%d)/%s?charset=utf8&parseTime=True"
-		conn = fmt.Sprintf(conn,
-			dbConfig.Username,
-			dbConfig.Password,
-			"tcp",
-			dbConfig.URL,
-			dbConfig.Port,
-			dbConfig.DbName)
-		DB, err = gorm.Open(mysql.New(mysql.Config{
-			DSN: conn,
-		}), &gorm.Config{})
-	case "sqlite3":
-		// see
-		// https://github.com/mattn/go-sqlite3/issues/677#issuecomment-450203752
-		DB, err = gorm.Open(sqlite.Open("file::memory:?cache=shared"), &gorm.Config{
-			// TODO make logger level configurable
-			Logger: logger.Default.LogMode(logger.Silent),
-		})
-		// see https://github.com/mattn/go-sqlite3/issues/209
-		maxDbConnections = 1
-	default:
-		errorString := "Cannot connect. Unsupported DB type: (" + dbConfig.DbType + ")"
-		log.Println(errorString)
-		return nil, errors.New(errorString)
-	}
-	if err != nil {
-		log.Println("Error!" + err.Error())
-		return nil, err
-	}
-
-	sqlDB, err := DB.DB()
-	if err != nil {
-		return nil, err
-	}
-
-	if err = sqlDB.Ping(); err != nil {
-		log.Println("Unable to verify connection to database")
-		return nil, err
-	}
-
-	sqlDB.SetMaxOpenConns(maxDbConnections)
-
-	return DB, nil
-}
