+++ main.go
@@ -8,9 +8,12 @@ import (
 	"os/signal"
 
 	"code.cloudfoundry.org/brokerapi/v13"
+	"github.com/cloud-gov/aws-broker/base"
 	"github.com/cloud-gov/aws-broker/catalog"
 	"github.com/cloud-gov/aws-broker/config"
+	"github.com/cloud-gov/aws-broker/services/elasticsearch"
 	"github.com/cloud-gov/aws-broker/services/rds"
+	"github.com/cloud-gov/aws-broker/services/redis"
 	brokertags "github.com/cloud-gov/go-broker-tags"
 	"github.com/riverqueue/river"
 
@@ -41,11 +44,19 @@ func run(ctx context.Context, out io.Writer) error {
 	// Create a new logger with the Text handler
 	logger := slog.New(handler)
 
-	logger.Debug("run :initializing database")
-	db, err := db.InternalDBInit(settings.DbConfig)
+	logger.Debug("run: initializing database")
+	db, err := db.DBInit(settings.DbConfig)
 	if err != nil {
 		return fmt.Errorf("error initializing database: %s", err)
 	}
+
+	logger.Debug("run: Migrating GORM models")
+	// Automigrate!
+	err = db.AutoMigrate(&rds.RDSInstance{}, &redis.RedisInstance{}, &elasticsearch.ElasticsearchInstance{}, &base.Instance{}, &jobs.AsyncJobMsg{}) // Add all your models here to help setup the database tables
+	if err != nil {
+		return fmt.Errorf("error migrating GORM models: %s", err)
+	}
+	logger.Debug("run: Migrated GORM models")
 
 	logger.Debug("run: initializing River workers and client")
 	workers := river.NewWorkers()
